refactor: getAttribute is replaced by classList in the menu test suite
getAttribute returns a string of classes, and using the toContain on
string makes expectation not reliable. If in future anything is added
in the end of class name, it can cause the functionality to break without
failing the test. So in the current implementation classList is used instead
of getAttribute. classList returns an array of classes on an element.
Using toContain on an array would make the test would fail even if
something is added at the end of a class name.